+++ src/index.js
@@ -1,0 +1,18 @@
+
+import BarSide from '@/components/barside/BarSide.vue';
+import DataTable from '@/components/table/DataTable.vue';
+import TextInput from '@/components/input/TextInput.vue';
+import TextInputArea from '@/components/input/TextInputArea.vue';
+
+
+const GtsPrint = {
+  install(app) {
+    app.component('BarSide', BarSide);
+    app.component('DataTable', DataTable);
+    app.component('TextInput', TextInput);
+    app.component('TextInputArea', TextInputArea);
+  }
+};
+
+ 
+export default GtsPrint;